hkcteyqjxm/mutilable-classification/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
appnope==0.1.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
asttokens==2.4.1
attrs==23.2.0
azure-core==1.30.1
azure-storage-blob==12.19.1
backoff==2.2.1
beautifulsoup4==4.12.3
blinker==1.7.0
botocore==1.34.93
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
coloredlogs==15.0.1
comm==0.2.2
contourpy==1.2.1
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.4
debugpy==1.8.1
decorator==5.1.1
deepdiff==7.0.1
Deprecated==1.2.14
emoji==2.12.1
entrypoints==0.4
environs==9.5.0
et-xmlfile==1.1.0
executing==2.0.1
filelock==3.13.4
filetype==1.2.0
Flask==3.0.3
flatbuffers==24.3.25
fonttools==4.52.1
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
grpcio==1.60.0
htbuilder==0.6.2
huggingface-hub==0.22.2
humanfriendly==10.0
ibm-cos-sdk==2.13.4
ibm-cos-sdk-core==2.13.4
ibm-cos-sdk-s3transfer==2.13.4
ibm_watson_machine_learning==1.0.356
ibm_watsonx_ai==0.2.6
idna==3.7
importlib_metadata==7.1.0
InstructorEmbedding==1.0.1
iopath==0.1.10
ipykernel==6.29.4
ipython==8.24.0
isodate==0.6.1
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
jmespath==1.0.1
joblib==1.4.0
jsonformer==0.12.0
jsonpatch==1.33
jsonpath-python==1.0.6
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain
langchain-community
langchain-core
langchain-experimental
langchain-ibm
langchain-text-splitters
langdetect
langsmith
layoutparser==0.3.4
lomond==0.3.3
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
matplotlib==3.9.0
matplotlib-inline==0.1.7
mdurl==0.1.2
minio==7.2.6
more-itertools==10.2.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.3
nltk==3.8.1
numpy==1.26.4
onnx==1.16.1
onnxruntime==1.18.0
opencv-python==4.9.0.80
openpyxl==3.1.2
ordered-set==4.1.0
orjson==3.10.1
packaging==23.2
pandas==2.1.4
parso==0.8.4
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.0
pexpect==4.9.0
pikepdf==8.15.1
pillow==10.3.0
pillow_heif==0.16.0
platformdirs==4.2.1
portalocker==2.8.2
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==16.0.0
pycparser==2.22
pycryptodome==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
pydeck==0.9.0b1
Pygments==2.17.2
pymilvus==2.4.0
pyparsing==3.1.2
pypdf==4.2.0
pypdfium2==4.30.0
pytesseract==0.3.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.0.1
python-iso639==2024.4.27
python-magic==0.4.27
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
rapidfuzz==3.9.1
referencing==0.35.0
regex==2024.4.16
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
safetensors==0.4.3
scikit-learn==1.4.2
scipy==1.13.0
sentence-transformers==2.7.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.29
st-annotated-text==4.0.1
stack-data==0.6.3
streamlit==1.33.0
streamlit-aggrid==1.0.5
streamlit-tags==1.2.8
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.4.0
threadpoolctl==3.4.0
timm==1.0.3
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.2.2
torchvision==0.17.2
tornado==6.4
tqdm==4.66.2
traitlets==5.14.3
transformers==4.40.1
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
ujson==5.9.0
unstructured==0.14.2
unstructured-client==0.22.0
unstructured-inference==0.7.33
unstructured.pytesseract==0.3.12
urllib3==2.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
